Workable draft of MultiTubeOD
Basic app to interact with one or more MultiTubeOD devices.
Configure Hardware is useful for when multiple devices are connected.
Start New Run helps the user collect blanks and set a reference tube.
Experiments run in independent infinite loops.
Run info and instrument usage are each stored in pickles.
The home page makes an accordion of active experiments where growth can be monitored and the run can be stopped.