only free up reference if not being shared
previously, every run set reference port usage to 0 when experiment stopped. This creates conflict with other running experiments that rely on that port. Added a condition to check if the port is being used by other running experiments and only set usage to 0 if that is the case.
Also, minor modifications to make code clearer.