Offload work onto objects
Created classes for
-Experiments
-Devices
-Ports

All are linked. Each class carries methods for handling the objects. Still don't have calibration figured out...

Objects are used for the interface, to collect user input and to build a header for the output file. The lightweight "timecourse.py" function is called as a subprocess and interprets the header to begin an assay.

No longer need the "sampling" module, since functions were either made irrelevant with or moved into appropriate class/static methods